+++ lib/src/gpt_plugin.dart
@@ -3,7 +3,6 @@ library gpt_plugins;
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
-import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:http/http.dart' as http;
@@ -12,9 +11,7 @@ import 'io/io.dart';
 import 'prompts.dart';
 
 part "plugins/batch_plugin.dart";
-
 part "plugins/experiment_plugin.dart";
-
 part "plugins/image_plugin.dart";
 
 abstract class GptPlugin {
@@ -31,6 +28,7 @@ abstract class GptPlugin {
   late String blockId;
   late String pluginName;
   late int blockRuns;
+  late String metricsFile;
   var currentBlockRun = 0;
 
   GptPlugin(this._projectConfig, this.block, this.io) {
@@ -43,6 +41,7 @@ abstract class GptPlugin {
     blockId = block["blockId"];
     pluginName = block["pluginName"];
     blockDataDir = "$dataDir/$blockId";
+    metricsFile = "$reportDir/metrics-$blockId.csv";
   }
 
   num apiCallCount() {
@@ -56,7 +55,7 @@ abstract class GptPlugin {
   Future<num> apiCallCountForBlock() async {
     num result = 0;
     final pluginConfiguration = block["configuration"];
-    final blockRuns = pluginConfiguration["blockRuns"] ?? 1;
+    final blockRuns = block["blockRuns"] ?? 1;
     for (var blockRun = 1; blockRun <= blockRuns; blockRun++) {
       final executions = block["executions"];
       for (var i = 0; i < executions.length; i++) {
@@ -73,7 +72,7 @@ abstract class GptPlugin {
     print("BlockId: $blockId, PluginName: $pluginName");
     final startTime = DateTime.now();
     final pluginConfiguration = block["configuration"];
-    blockRuns = pluginConfiguration["blockRuns"] ?? 1;
+    blockRuns = block["blockRuns"] ?? 1;
     createDirectoryIfNotExist(blockDataDir);
     final blockResults = [];
     for (var blockRun = 1; blockRun <= blockRuns; blockRun++) {
@@ -148,7 +147,7 @@ abstract class GptPlugin {
   }
 
   Future<Map<String, dynamic>> sendHttpPostRequest(
-      requestBody, urlPath, dryRun) async {
+      requestBody, urlPath, executionId, tag, dryRun) async {
     final requestBodyStr = jsonEncode(requestBody);
     if (dryRun) {
       print("\tPOST to https://api.openai.com/$urlPath");
@@ -172,6 +171,7 @@ abstract class GptPlugin {
       } else {
         print(
             '\t\tOpenAI Request failed with status code: ${response.statusCode}');
+        print(requestBodyStr);
         logFailedRequest(requestBodyStr);
         return {"errorCode": response.statusCode};
       }
@@ -179,6 +179,7 @@ abstract class GptPlugin {
           jsonDecode(await readResponse(response));
       responseBody.addAll({"requestTime": (endTime - startTime)});
       logRequestAndResponse(requestBodyStr, responseBody);
+      writeMetrics(responseBody, executionId, tag);
       return responseBody;
     } catch (e) {
       print('Error occurred during the request: $e');
@@ -186,7 +187,37 @@ abstract class GptPlugin {
     return {};
   }
 
+  Future<Map<String, dynamic>> readJsonFile(String filePath) async {
+    String jsonString = await io.readFileAsString(filePath);
+    return jsonDecode(jsonString);
+  }
+
   Future<String> readResponse(HttpClientResponse response) async {
     return response.transform(utf8.decoder).join();
+  }
+
+  Future<void> writeMetrics(responseBody, executionId, tag) async {
+    final responseId = responseBody["id"] ?? "N/A";
+    final usage = responseBody["usage"];
+    final promptTokens = usage?['prompt_tokens'] ?? 0;
+    final completionTokens = usage?['completion_tokens'] ?? 0;
+    final totalTokens = usage?['total_tokens'] ?? 0;
+
+    final file = File(metricsFile);
+    bool exists = await file.exists();
+    if (!exists) {
+      await file.writeAsString(
+          "request_id, executionId, tag, request_time, prompt_tokens, completion_tokens, total_tokens\n");
+    }
+    final sink = File(metricsFile).openWrite(mode: FileMode.append);
+    try {
+      final requestTime = responseBody['requestTime'];
+      sink.write(
+          "$responseId, $executionId, $tag, $requestTime, $promptTokens, $completionTokens, $totalTokens\n");
+    } catch (e) {
+      throw Exception('Error occurred while writing file: $e');
+    } finally {
+      sink.close();
+    }
   }
 }